fix: validation issue added: add all input fields to output fields
