+++ src/main/java/io/cdap/plugin/WindowAggregation.java
@@ -13,9 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-
 package io.cdap.plugin;
-
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Strings;
@@ -24,7 +22,6 @@ import io.cdap.cdap.api.annotation.Name;
 import io.cdap.cdap.api.annotation.Plugin;
 import io.cdap.cdap.api.data.format.StructuredRecord;
 import io.cdap.cdap.api.data.schema.Schema;
-import io.cdap.cdap.api.spark.sql.DataFrames;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.cdap.etl.api.PipelineConfigurer;
 import io.cdap.cdap.etl.api.StageConfigurer;
@@ -33,20 +30,10 @@ import io.cdap.cdap.etl.api.batch.SparkExecutionPluginContext;
 import io.cdap.cdap.etl.api.batch.SparkPluginContext;
 import io.cdap.cdap.etl.api.lineage.field.FieldOperation;
 import io.cdap.cdap.etl.api.lineage.field.FieldTransformOperation;
-import io.cdap.plugin.function.DiscretePercentile;
 import org.apache.spark.api.java.JavaRDD;
-import org.apache.spark.api.java.function.Function;
-import org.apache.spark.sql.Column;
-import org.apache.spark.sql.Dataset;
-import org.apache.spark.sql.Row;
-import org.apache.spark.sql.SQLContext;
-import org.apache.spark.sql.expressions.Window;
-import org.apache.spark.sql.expressions.WindowSpec;
-import org.apache.spark.sql.functions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.security.InvalidParameterException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedList;
@@ -77,32 +64,34 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
   public void configurePipeline(PipelineConfigurer pipelineConfigurer) {
 
     StageConfigurer stageConfigurer = pipelineConfigurer.getStageConfigurer();
+    FailureCollector failureCollector = stageConfigurer.getFailureCollector();
     Schema inputSchema = stageConfigurer.getInputSchema();
 
     List<String> partitionFields = config.getPartitionFields();
-    List<WindowAggregationConfig.FunctionInfo> aggregates = config.getAggregates();
+    List<WindowAggregationConfig.FunctionInfo> aggregates = config.getAggregates(failureCollector);
     List<String> partitionOrderFields = config.getPartitionOrderFields();
 
     if (inputSchema == null || partitionFields.isEmpty() || aggregates.isEmpty()) {
+      failureCollector.getOrThrowException();
       stageConfigurer.setOutputSchema(null);
       return;
     }
 
-    validate(inputSchema, partitionFields, aggregates, partitionOrderFields, stageConfigurer.getFailureCollector());
-    stageConfigurer.getFailureCollector().getOrThrowException();
+    validate(inputSchema, partitionFields, aggregates, partitionOrderFields, failureCollector);
+    failureCollector.getOrThrowException();
     stageConfigurer.setOutputSchema(getOutputSchema(inputSchema, aggregates));
   }
 
   private void validate(Schema inputSchema, List<String> partitionFields,
-                       List<WindowAggregationConfig.FunctionInfo> aggregates, List<String> partitionOrderFields,
-                       FailureCollector collector) {
+                        List<WindowAggregationConfig.FunctionInfo> aggregates, List<String> partitionOrderFields,
+                        FailureCollector collector) {
 
     for (String partitionField : partitionFields) {
       Schema.Field field = inputSchema.getField(partitionField);
       if (field == null) {
         collector.addFailure(String.format("Partition field '%s' must exist in input schema.", partitionField),
                              "")
-          .withConfigElement("partitionFields", partitionField);
+          .withConfigElement(WindowAggregationConfig.NAME_PARTITION_FIELD, partitionField);
       }
     }
 
@@ -118,7 +107,7 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
         collector.addFailure(
           String.format("Invalid aggregate %s: Field '%s' does not exist in input schema.", functionInfo.description(),
                         functionFieldName), String.format("Field '%s' must exist in input schema.", functionFieldName))
-          .withConfigProperty("aggregates");
+          .withConfigProperty(WindowAggregationConfig.NAME_AGGREGATES);
       } else {
         inputFieldSchema = inputField.getSchema();
       }
@@ -153,13 +142,17 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
         collector.addFailure(
           String.format("Can not order by field '%s' because it does not exist in input schema", partitionOrderField),
           "Order field '%s' should exist must exist in input schema.").
-          withConfigProperty("partitionFieldsOrder");
+          withConfigProperty(WindowAggregationConfig.NAME_PARTITION_ORDER);
       }
     }
   }
 
   private Schema getOutputSchema(Schema inputSchema, List<WindowAggregationConfig.FunctionInfo> aggregates) {
     List<Schema.Field> outputFields = new ArrayList<>(aggregates.size());
+    List<Schema.Field> inputSchemaFields = inputSchema.getFields();
+    if (inputSchemaFields != null) {
+      outputFields.addAll(inputSchemaFields);
+    }
     for (WindowAggregationConfig.FunctionInfo aggregate : aggregates) {
       outputFields.add(Schema.Field.of(aggregate.getAlias(), aggregate.getFunction().getOutputSchema()));
     }
@@ -169,18 +162,24 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
   @Override
   public void prepareRun(SparkPluginContext context) throws Exception {
     super.prepareRun(context);
-    validate(context.getInputSchema(), config.getPartitionFields(), config.getAggregates(),
-             config.getPartitionOrderFields(), context.getFailureCollector());
-    Schema outputSchema = getOutputSchema(context.getInputSchema(), config.getAggregates());
+    FailureCollector failureCollector = context.getFailureCollector();
+    List<WindowAggregationConfig.FunctionInfo> aggregates = config.getAggregates(failureCollector);
+    validate(context.getInputSchema(), config.getPartitionFields(), aggregates, config.getPartitionOrderFields(),
+             failureCollector);
+    failureCollector.getOrThrowException();
+    Schema outputSchema = getOutputSchema(context.getInputSchema(), aggregates);
     recordLineage(context, outputSchema);
   }
 
   @Override
   public void initialize(SparkExecutionPluginContext context) throws Exception {
     super.initialize(context);
-    validate(context.getInputSchema(), config.getPartitionFields(), config.getAggregates(),
-             config.getPartitionOrderFields(), context.getFailureCollector());
-    outputSchema = getOutputSchema(context.getInputSchema(), config.getAggregates());
+    FailureCollector failureCollector = context.getFailureCollector();
+    List<WindowAggregationConfig.FunctionInfo> aggregates = config.getAggregates(failureCollector);
+    validate(context.getInputSchema(), config.getPartitionFields(), aggregates, config.getPartitionOrderFields(),
+             failureCollector);
+    failureCollector.getOrThrowException();
+    outputSchema = getOutputSchema(context.getInputSchema(), aggregates);
   }
 
   private void recordLineage(SparkPluginContext context, Schema schema) {
@@ -202,7 +201,7 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
 
     LinkedList<FieldOperation> fllOperations = new LinkedList<>();
     // for every function record the field level operation details
-    for (WindowAggregationConfig.FunctionInfo functionInfo : config.getAggregates()) {
+    for (WindowAggregationConfig.FunctionInfo functionInfo : config.getAggregates(context.getFailureCollector())) {
 
       String functionDescription = generateFunctionDescription(functionInfo);
       String description = String.format("Generated field '%s' by partitioning on the column%s '%s' %s %s %s",
@@ -259,186 +258,6 @@ public class WindowAggregation extends SparkCompute<StructuredRecord, Structured
     if (inputSchema == null) {
       throw new Exception("Input schema is null. Input schema can not be null at this stage");
     }
-    JavaRDD<Row> map = javaRDD.map(
-      structuredRecord -> DataFrames.toRow(structuredRecord, DataFrames.toDataType(inputSchema)));
-
-    SQLContext sqlContext = new SQLContext(sparkExecutionPluginContext.getSparkContext());
-
-
-    Dataset<Row> data = sqlContext.createDataFrame(map, DataFrames.toDataType(inputSchema));
-    WindowSpec spec = Window.partitionBy(config.getPartitionsColumns()).orderBy(config.getPartitionOrderColumns());
-
-    switch (config.getWindowFrameType()) {
-      case ROW:
-        spec = spec.rowsBetween(config.getFrameDefinitionPrecedingBound(),
-                                config.getFrameDefinitionFollowingBound());
-        break;
-      case RANGE:
-        spec = spec.rangeBetween(config.getFrameDefinitionPrecedingBound(),
-                                 config.getFrameDefinitionFollowingBound());
-        break;
-      case NONE:
-      default:
-    }
-
-    List<WindowAggregationConfig.FunctionInfo> aggregatesData = config.getAggregates();
-
-
-    for (WindowAggregationConfig.FunctionInfo aggregatesDatum : aggregatesData) {
-      if (aggregatesDatum.getFunction() == WindowAggregationConfig.Function.DISCRETE_PERCENTILE) {
-        data = applyCustomFunction(sqlContext, aggregatesDatum, data, spec, inputSchema);
-      } else {
-        data = apply(aggregatesDatum, data, spec);
-      }
-    }
-
-    JavaRDD<StructuredRecord> rowJavaRDD = data.javaRDD().map(new RowToRecord(outputSchema));
-
-    String numberOfPartitionsString = config.getNumberOfPartitions();
-
-    if (Strings.isNullOrEmpty(numberOfPartitionsString)) {
-      return sparkExecutionPluginContext.getSparkContext()
-        .parallelize(rowJavaRDD.collect());
-    }
-
-    return sparkExecutionPluginContext.getSparkContext().parallelize(rowJavaRDD.collect(),
-                                                                     Integer.parseInt(numberOfPartitionsString));
-  }
-
-  private Dataset<Row> apply(WindowAggregationConfig.FunctionInfo data, Dataset<Row> dataFrame,
-                             WindowSpec spec) {
-    Column aggregateColumn = getAggregateColumn(data).over(spec);
-    return dataFrame.withColumn(data.getAlias(), aggregateColumn);
-  }
-
-  private Dataset<Row> applyCustomFunction(SQLContext sqlContext, WindowAggregationConfig.FunctionInfo aggregatesDatum,
-                                           Dataset<Row> dataFrame, WindowSpec spec, Schema inputSchema) {
-    String[] args = aggregatesDatum.getArgs();
-    if (args.length < 1) {
-      throw new InvalidParameterException("Discrete Percentile must have at least 1 arguments");
-    }
-
-    String percentileString = args[0];
-    double percentile;
-    try {
-      percentile = Double.parseDouble(percentileString);
-    } catch (NumberFormatException e) {
-      throw new InvalidParameterException("Discrete Percentile, first argument must be double value");
-    }
-
-    if (percentile > 1 || percentile < 0) {
-      throw new InvalidParameterException("Discrete Percentile, percentile must be in range 0.0-1.0");
-    }
-
-    Schema discretePercentileInputSchema = schemaForInputField(inputSchema, aggregatesDatum.getFieldName());
-    DiscretePercentile discretePercentile = new DiscretePercentile(percentile, discretePercentileInputSchema);
-    sqlContext.udf().register("DISCRETE_PERCENTILE", discretePercentile);
-
-    return apply(aggregatesDatum, dataFrame, spec);
-  }
-
-  private Column getAggregateColumn(WindowAggregationConfig.FunctionInfo data) {
-    String[] args = data.getArgs();
-    switch (data.getFunction()) {
-      case RANK:
-        return functions.rank();
-      case DENSE_RANK:
-        return functions.dense_rank();
-      case PERCENT_RANK:
-        return functions.percent_rank();
-      case N_TILE:
-        checkFunctionArguments(args, 1, "N tile");
-        return functions.ntile(Integer.parseInt(args[0]));
-      case ROW_NUMBER:
-        return functions.row_number();
-      case MEDIAN:
-        return functions.callUDF("percentile", functions.col(data.getFieldName()), functions.lit(0.5));
-      case CONTINUOUS_PERCENTILE: {
-        checkFunctionArguments(args, 1, "Continuous Percentile");
-        String arg = args[0];
-        Column lit = functions.lit(arg);
-        return functions.callUDF("percentile", functions.col(data.getFieldName()), lit);
-      }
-      case DISCRETE_PERCENTILE: {
-        return functions.callUDF("DISCRETE_PERCENTILE", functions.col(data.getFieldName()));
-      }
-      case LEAD: {
-        checkFunctionArguments(args, 1, "LEAD");
-        return functions.lead(data.getFieldName(), Integer.parseInt(args[0]));
-      }
-      case LAG:
-        checkFunctionArguments(args, 1, "LAG");
-        return functions.lag(data.getFieldName(), Integer.parseInt(args[0]));
-      case FIRST: {
-        boolean ignoreNull = data.isIgnoreNull();
-        if (args.length != 0) {
-          ignoreNull = ignoreNull || Boolean.parseBoolean(args[0]);
-        }
-        return functions.first(data.getFieldName(), ignoreNull);
-      }
-      case LAST: {
-        boolean ignoreNull = data.isIgnoreNull();
-        if (args.length != 0) {
-          ignoreNull = ignoreNull || Boolean.parseBoolean(args[0]);
-        }
-        return functions.last(data.getFieldName(), ignoreNull);
-      }
-      case CUMULATIVE_DISTRIBUTION:
-        return functions.cume_dist();
-      case ACCUMULATE:
-        return functions.sum(data.getFieldName());
-      default:
-        throw new InvalidParameterException("Can not find function " + data.getFieldName());
-    }
-  }
-
-  private void checkFunctionArguments(String[] args, int minimumArguments, String function) {
-    if (args.length < minimumArguments) {
-      throw new InvalidParameterException(
-        String.format("%s must have at least %s arguments", function, minimumArguments));
-    }
-  }
-
-  private Schema schemaForInputField(Schema inputSchema, String inputFieldName) {
-    if (inputSchema == null) {
-      return null;
-    }
-
-    Schema.Field schemaField = inputSchema.getField(inputFieldName);
-    if (schemaField == null) {
-      return null;
-    }
-
-    return schemaField.getSchema();
-  }
-
-  /**
-   * Function to map from {@link Row} to {@link StructuredRecord}.
-   */
-  public static final class RowToRecord implements
-    Function<Row, StructuredRecord> {
-
-    private final Schema outputSchema;
-
-    public RowToRecord(Schema outputSchema) {
-      this.outputSchema = outputSchema;
-    }
-
-    @Override
-    public StructuredRecord call(Row row) throws Exception {
-
-      Schema rowSchema = DataFrames.toSchema(row.schema());
-      StructuredRecord structuredRecord = DataFrames.fromRow(row, rowSchema);
-
-      if (outputSchema == null || outputSchema.getFields() == null) {
-        throw new Exception("Invalid output schema");
-      }
-
-      StructuredRecord.Builder builder = StructuredRecord.builder(outputSchema);
-      for (Schema.Field field : outputSchema.getFields()) {
-        builder.set(field.getName(), structuredRecord.get(field.getName()));
-      }
-      return builder.build();
-    }
+    return WindowsAggregationUtil.transform(sparkExecutionPluginContext, javaRDD, config, inputSchema, outputSchema);
   }
 }
